backend/rag_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pinecone-client==3.0.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
llama-cpp-python==0.2.11 